Added flash messages

+++ app/controllers/votes_controller.rb
@@ -6,18 +6,24 @@ class VotesController < ApplicationController
 
     user_votes = @votable.votes.where(user_id: helpers.current_user.id)
 
-    if user_votes.exists?
-      user_votes.first.destroy
-    end
+    user_votes.first.destroy if user_votes.exists?
 
-    @vote.save
+    if @vote.save
+      flash[:success] = "#{@votable.class.name} #{@vote.vote_type}d successfully"
+    else
+      flash[:warning] = "Failed to #{@vote.vote_type} #{@votable.class.name.downcase}"
+    end
 
     redirect_to @votable.get_question
   end
 
   def destroy
     @vote = Vote.find(params[:id])
-    @vote.destroy
+    if @vote.destroy
+      flash[:success] = "#{@vote.vote_type.capitalize} deleted successfully"
+    else
+      flash[:warning] = "Failed to delete #{@vote.vote_type}"
+    end
 
     redirect_to @vote.content.get_question
   end
